Fix captcha error on loadballanced production
by tolarating any failure with no http status and response text resulting from that call.
The captcha checking is on the success branch and is not affected and
nor the 400, 500 and any server error response.
(The XhrHttpRequest is cancelled by the browser itself before the final
response, because of mixed content, resulted by following an intermediary 302 response with wrong http proto)